fix long int, remove extra long

+++ 0x00-hello_world/6-size.c
@@ -10,7 +10,7 @@ int main(void)
 {
 	char typeChar;
 	int typeInt;
-	long long int typeLong;
+	long int typeLong;
 	long long int typeLongLong;
 	float typeFloat;
